Crash trying to drag the entire text of a long book out of Mail compose view https://bugs.webkit.org/show_bug.cgi?id=173042 <rdar://problem/32382059>
Reviewed by Simon Fraser.

Source/WebCore:

Creating a TextIndicator that is hundreds of thousands of pixels tall
is not a good idea. We introduced TextIndicatorOptionDoNotClipToVisibleRect
so that (for example) an image that overhangs the edge of the page would
contain the entire image, instead of just the visible fragment.
Instead, rename this option to ExpandClipBeyondVisibleRect, and make it
not unclip *entirely*, but expand the clip significantly (50% in each direction)
beyond the bounds of the current visible rect. Also, make use of exposed
area information for this clipping; otherwise, clients with very large views
inside scrollable areas (like Mail) would still try to create large TextIndicators.

* page/TextIndicator.cpp:
(WebCore::initializeIndicator):
* page/TextIndicator.h:

Source/WebKit2:

* WebProcess/WebPage/ios/FindControllerIOS.mm:
* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::didConcludeEditDataInteraction):
Adapt to the new name.


Identifier: 189895@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217870 268f45cc-cd09-0410-ab3c-d52691b4dbfc