[JSC] Shrink Structure size https://bugs.webkit.org/show_bug.cgi?id=173239
Reviewed by Mark Lam.

We find that the size of our Structure is slightly enlarged due to paddings.
By changing the order of members, we can reduce the size from 120 to 112.
This is good because 120 and 112 are categorized into different size classes.
For 120, we allocate 128 bytes. And for 112, we allocate 112 bytes.
We now save 16 bytes per Structure for free.

* runtime/ConcurrentJSLock.h:
* runtime/Structure.cpp:
(JSC::Structure::Structure):
* runtime/Structure.h:


Identifier: 190082@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218070 268f45cc-cd09-0410-ab3c-d52691b4dbfc