[Win] WebKit fails to paint the bottom lines of the window at 125% scale. https://bugs.webkit.org/show_bug.cgi?id=173318
Reviewed by Alex Christensen.

The logical frame size is currently calculated by scaling the frame size in pixels,
and rounding the coordinates to the nearest integers. The computed logical frame
size will in some cases be smaller than the scaled frame size in pixels, since it is
represented as an IntSize. In these cases, the bottom line will not be painted.
Instead of calculating the logical frame size by rounding the scaled frame in
pixels, find the enclosing rectangle of the scaled frame rectangle in pixels.

* WebView.cpp:
(WebView::sizeChanged):


Identifier: 190180@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218186 268f45cc-cd09-0410-ab3c-d52691b4dbfc