Rebaseline media/modern-media-controls/fullscreen-support https://bugs.webkit.org/show_bug.cgi?id=173199
Reviewed by Simon Fraser.

Source/WebCore:

Test: media/modern-media-controls/fullscreen-support/fullscreen-support-press.html

We no longer need to track the "error" event for FullscreenSupport since we don't actually
enable or disable the fullscreen button when an error happens, instead the media controls
code will simply not present it since we'll be showing a placard instead.

* Modules/modern-media-controls/media/fullscreen-support.js:
(FullscreenSupport.prototype.get mediaEvents):

LayoutTests:

We no longer need to test the "error" event for FullscreenSupport since we don't actually
enable or disable the fullscreen button when an error happens, instead the media controls
code will simply not present it since we'll be showing a placard instead.

We also remove the iPad-specific test since the fullscreen button is now accessible on an
iPhone as well, so it no longer needs a special iPad test.

* media/modern-media-controls/fullscreen-support/fullscreen-support-enabled-expected.txt:
* media/modern-media-controls/fullscreen-support/fullscreen-support-enabled.html:
* media/modern-media-controls/fullscreen-support/fullscreen-support-press-expected.txt: Renamed from LayoutTests/media/modern-media-controls/fullscreen-support/fullscreen-support-click-expected.txt.
* media/modern-media-controls/fullscreen-support/fullscreen-support-press.html: Renamed from LayoutTests/media/modern-media-controls/fullscreen-support/fullscreen-support-click.html.
* media/modern-media-controls/fullscreen-support/ipad/fullscreen-support-tap-expected.txt: Removed.
* media/modern-media-controls/fullscreen-support/ipad/fullscreen-support-tap.html: Removed.
* platform/ios-simulator/TestExpectations:
* platform/mac/TestExpectations:

Identifier: 190207@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218231 268f45cc-cd09-0410-ab3c-d52691b4dbfc