[GTK][WPE] Use CMAKE_THREAD_LIBS_INIT in WTF platform CMake files https://bugs.webkit.org/show_bug.cgi?id=173310
Reviewed by Xabier Rodriguez-Calvar.

In the WTF platform CMake files for the GTK+ and WPE ports, use the
CMAKE_THREAD_LIBS_INIT variable to specify the threading library we
should be linking against. On some platforms this variable will be
empty because the given compiler will automatically handle the -pthread
compiler flag and sort out linking on its own.

The same approach is already used for the JSCOnly configuration.

* wtf/PlatformGTK.cmake:
* wtf/PlatformWPE.cmake:


Identifier: 190221@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218245 268f45cc-cd09-0410-ab3c-d52691b4dbfc