Ensure we only log changes to the ScriptedAnimationController suspended state in debug builds https://bugs.webkit.org/show_bug.cgi?id=173423
Patch by Antoine Quint <graouts@apple.com> on 2017-06-15
Reviewed by Tim Horton.

We added logging for when the suspended state of the scripted animation controller would change in
webkit.org/b/173326. It was meant to only be enabled in debug builds and we actually did the wrong
thing and enabled it in non-debug builds.

We also added a setting that wasn't used and that we are removing here.

* dom/ScriptedAnimationController.cpp:
(WebCore::ScriptedAnimationController::logSuspendCount):
* page/Settings.in:
* testing/InternalSettings.cpp:
(WebCore::InternalSettings::resetToConsistentState):
(WebCore::InternalSettings::shouldLogScritedAnimationControllerSuspensionChange): Deleted.
(WebCore::InternalSettings::setShouldLogScritedAnimationControllerSuspensionChange): Deleted.
* testing/InternalSettings.h:
* testing/InternalSettings.idl:

Identifier: 190307@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218351 268f45cc-cd09-0410-ab3c-d52691b4dbfc