REGRESSION: AirPlay button is incorrectly highlighted in inline and fullscreen https://bugs.webkit.org/show_bug.cgi?id=173446
Patch by Antoine Quint <graouts@apple.com> on 2017-06-15
Reviewed by Dean Jackson.

Source/WebCore:

A button's color should be applied to its <picture> element, not the containing
<button> element.

* Modules/modern-media-controls/controls/airplay-button.css:
(button.airplay.on > picture):
(button.airplay.on): Deleted.

LayoutTests:

Update the test to explicitly check the color on the <picture>. We also make the test check
the color with a value depending on the platform, so we can use a single expectation file.

* media/modern-media-controls/airplay-button/airplay-button-on-expected.txt:
* media/modern-media-controls/airplay-button/airplay-button-on.html:
* platform/ios/media/modern-media-controls/airplay-button/airplay-button-on-expected.txt: Removed.

Identifier: 190328@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218377 268f45cc-cd09-0410-ab3c-d52691b4dbfc