Don't use WebCore Timer from code that runs in the UI process. https://bugs.webkit.org/show_bug.cgi?id=173460 rdar://problem/32750731
Patch by Jeremy Jones <jeremyj@apple.com> on 2017-06-16
Reviewed by Jer Noble.

If a WebCore Timer is fired from the WebKit2 UI process, it will
create a web thread, which can cause a crash in this instance.

It was also causing the timer to be called back on the wrong thread.

* platform/ios/WebVideoFullscreenInterfaceAVKit.h:
* platform/ios/WebVideoFullscreenInterfaceAVKit.mm:
(WebVideoFullscreenInterfaceAVKit::WebVideoFullscreenInterfaceAVKit):

Identifier: 190358@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218407 268f45cc-cd09-0410-ab3c-d52691b4dbfc